Check if Processor running before shutdown

+++ stop_connect.sh
@@ -11,12 +11,17 @@ if [ "$CONDA_DEFAULT_ENV" == "proda" ] || [ "$CONDA_DEFAULT_ENV" == "deva" ]; th
     fi
 elif [ "$CONDA_DEFAULT_ENV" == "prodp" ] || [ "$CONDA_DEFAULT_ENV" == "devp" ]; then
     echo "Shutting down Connect Processor";
-    kill -SIGTERM `cat pid.log`;
-    sleep 10;
-    while [ $(ps -e | grep -c $(cat pid.log)) -gt 0 ]; do
-        echo "Connect Processor still running";
-        sleep 60;
-    done
-    echo "Connect Processor shut down";
+    if [ -f pid.log ]; then
+        kill -SIGTERM `cat pid.log`;
+        sleep 10;
+        while [ $(ps -e | grep -c $(cat pid.log)) -gt 0 ]; do
+            echo "Connect Processor still running";
+            sleep 60;
+        done
+        rm pid.log
+        echo "Connect Processor shut down";
+    else
+        echo "Connect Processor not running";
+    fi
 fi
